Re-engineering. Required changes to include files.

+++ AlertRig/src/fixstim/PlaidStimSet.h
@@ -1,10 +1,10 @@
 #pragma once
 #include "StimSet.h"
-
+#include "ARPlaidSpec.h"
 
 class PlaidStimSet : public StimSet
 {
-	std::vector<int> m_vecContrast;
+	std::vector<double> m_vecContrast;
 	std::vector<double> m_vecSF;
 	std::vector<double> m_vecTF;
 	std::vector<double> m_vecOriDegrees;
@@ -17,10 +17,8 @@ class PlaidStimSet : public StimSet
 	void getDriftPos(double t, double ppd, double tf, const alert::ARPlaidSubGr& gr1, const alert::ARPlaidSubGr& gr2, double& xpos, double& ypos);
 
 public:
-	PlaidStimSet(const ARGratingSpec& g0, const ARGratingSpec& g1, const std::vector<int> vecContrast, const std::vector<double> vecSF, const std::vector<double> vecTF, const std::vector<double>& oris);
-	PlaidStimSet(const ARGratingSpec& g0, const ARGratingSpec& g1, int contrast, double sf, double tf, double ori);
-	//PlaidStimSet(const std::vector<int> vecContrast, const std::vector<double> vecSF, const std::vector<double> vecTF, const std::vector<double>& oris);
-	//PlaidStimSet(int contrast, double sf, double tf, double ori);
+	PlaidStimSet(const ARGratingSpec& g0, const ARGratingSpec& g1, const std::vector<double> vecContrast, const std::vector<double> vecSF, const std::vector<double> vecTF, const std::vector<double>& oris);
+	PlaidStimSet(const ARGratingSpec& g0, const ARGratingSpec& g1, double contrast, double sf, double tf, double ori);
 
 	virtual ~PlaidStimSet() {};
